Intro - simple web server
Pretty simple; creates a webserver using the builtin node.http library.

Calls the newly defined createServer to create the server, and passes in
an anon function for handling the requests.

When a request comes in, we set the status code & set a header; can
presumably set an (infinite?) number of headers. The response has a
`.end` option (well, method if we're being precise) that we can write to
that I'm assuming finishes up the response; writes a `Content-Length`
header and dumps the entire text contents (or whatever) to the client.
Unsure if this is a buffered write or if it blocks.

For instance, if the client reads only 1 byte per second, and 60 chars
are added, does the server block for 60s, or is it smarter than that?